Token generated by parent organisation should be usuable by the child organisation
Started developing this feature. The files that are being worked on are OrganisationalUnitService. A new funciton named getChildren is used to return the current organisation and its children. The main plan is when the user inserts the organisation token and name when signing up, then the system should check if the token is made my the current organisation or one of its parents and if so then it should allow the token. More checks are needed.