hotfix: org list returns all organisations for invited user with agency domain
When a user has been manually invited to LPG, they should be returned the full list of organisations. This fixes a small bug where if a user was invited, but their email domain was also part of an agency token, they would be returned an empty list. This fix instead checks whether the user has an agencyTokenUid on their identity object, rather than checking by domain.

- Removed redundant test file
- Updated the default datasource value to point to csrs database to make local config consistant to rest of environments.